[CONFIG] Temporarily rename button, tooltip and navbar components
