+++ src/components/buttons/index.ts
@@ -1,2 +1,2 @@
-export { Button } from './button/Button';
+export { Button } from './button/button-temp';
 export { IconButton } from './icon-button/IconButton';
