+++ src/components/index.ts
@@ -2,4 +2,4 @@ export { ActionLoader } from './action-loader/ActionLoader';
 export { ProfilePic } from './profile-pic/ProfilePic';
 export { SuspenseLoader } from './suspense-loader/SuspenseLoader';
 export { Timestamp } from './timestamp/Timestamp';
-export { Tooltip } from './tooltip/Tooltip';
+export { Tooltip } from './tooltip/tooltip-temp';
